final commit SGLC

+++ brs_modules/brs_ovpn.js
@@ -168,9 +168,11 @@ ovpn.prototype.exportConfiguration = function(){
 			fs.writeFile(path+this.parent+".key",this.key,(error)=>{
 				if (error) return rej({cantWriteKey:error});
 			});
-			fs.writeFile(path+this.parent+"UP.sh",'#!/bin/bash\n',(error)=>{
-				if (error) return rej({cantWriteUpScript:error});
-			});
+			if (!fs.existsSync(path+this.parent+"UP.sh")) {
+				fs.writeFile(path+this.parent+"UP.sh",'#!/bin/bash\n',(error)=>{
+					if (error) return rej({cantWriteUpScript:error});
+				});
+			}
 			fs.writeFile(path+this.parent+".sh","#!/bin/bash\nmkdir /dev/net\nmknod /dev/net/tun c 10 200\nchmod 666 /dev/net/tun\n/usr/sbin/openvpn --config /share/"+this.parent+".ovpn > /var/log/opvn.log &\n\n",(error)=>{
 				if (error) return rej({cantWriteRcLocalScript:error});	
 			});
